wip: convert most modules
