+++ iroh-willow/src/lib.rs
@@ -3,13 +3,13 @@
 #![allow(missing_docs)]
 #![deny(unsafe_code)]
 
-// pub mod auth;
 // pub mod engine;
-// pub mod form;
+pub mod form;
 // pub mod net;
 pub mod proto;
 // pub mod session;
-// pub mod store;
+pub mod interest;
+pub mod store;
 pub mod util;
 
 /// To break symmetry, we refer to the peer that initiated the synchronisation session as Alfie,
