+++ iroh-willow/src/store/memory.rs
@@ -5,13 +5,13 @@ use std::rc::Rc;
 use anyhow::Result;
 
 use crate::{
-    auth::{CapSelector, CapabilityPack},
+    interest::{CapSelector, CapabilityPack},
     proto::{
+        data_model::{AuthorisedEntry, Entry, EntryExt, WriteCapability},
         grouping::{Range, RangeEnd, ThreeDRange},
-        keys::{NamespaceSecretKey, UserId, UserSecretKey},
-        meadowcap,
-        sync::{Fingerprint, ReadAuthorisation},
-        willow::{AuthorisedEntry, Entry, NamespaceId, WriteCapability},
+        keys::{NamespaceId, NamespaceSecretKey, UserId, UserSecretKey},
+        meadowcap::{self, is_wider_than, ReadAuthorisation},
+        wgps::Fingerprint,
     },
     store::traits::{self, RangeSplit, SplitAction, SplitOpts},
 };
@@ -126,44 +126,44 @@ impl traits::EntryReader for Rc<RefCell<EntryStore>> {
         let mid = entries.get(split_index).expect("not empty");
         let mut ranges = vec![];
         // split in two halves by subspace
-        if mid.subspace_id != range.subspaces.start {
+        if *mid.subspace_id() != range.subspaces().start {
             ranges.push(ThreeDRange::new(
-                Range::new(range.subspaces.start, RangeEnd::Closed(mid.subspace_id)),
-                range.paths.clone(),
-                range.times,
+                Range::new_closed(range.subspaces().start, *mid.subspace_id()).unwrap(),
+                range.paths().clone(),
+                *range.times(),
             ));
             ranges.push(ThreeDRange::new(
-                Range::new(mid.subspace_id, range.subspaces.end),
-                range.paths.clone(),
-                range.times,
+                Range::new(*mid.subspace_id(), range.subspaces().end),
+                range.paths().clone(),
+                *range.times(),
             ));
         }
         // split by path
-        else if mid.path != range.paths.start {
+        else if *mid.path() != range.paths().start {
             ranges.push(ThreeDRange::new(
-                range.subspaces,
+                *range.subspaces(),
                 Range::new(
-                    range.paths.start.clone(),
-                    RangeEnd::Closed(mid.path.clone()),
+                    range.paths().start.clone(),
+                    RangeEnd::Closed(mid.path().clone()),
                 ),
-                range.times,
+                *range.times(),
             ));
             ranges.push(ThreeDRange::new(
-                range.subspaces,
-                Range::new(mid.path.clone(), range.paths.end.clone()),
-                range.times,
+                *range.subspaces(),
+                Range::new(mid.path().clone(), range.paths().end.clone()),
+                *range.times(),
             ));
         // split by time
         } else {
             ranges.push(ThreeDRange::new(
-                range.subspaces,
-                range.paths.clone(),
-                Range::new(range.times.start, RangeEnd::Closed(mid.timestamp)),
+                *range.subspaces(),
+                range.paths().clone(),
+                Range::new(range.times().start, RangeEnd::Closed(mid.timestamp())),
             ));
             ranges.push(ThreeDRange::new(
-                range.subspaces,
-                range.paths.clone(),
-                Range::new(mid.timestamp, range.times.end),
+                *range.subspaces(),
+                range.paths().clone(),
+                Range::new(mid.timestamp(), range.times().end),
             ));
         }
         let mut out = vec![];
@@ -210,7 +210,7 @@ impl traits::EntryStorage for Rc<RefCell<EntryStore>> {
 
     fn ingest_entry(&self, entry: &AuthorisedEntry) -> Result<bool> {
         let mut slf = self.borrow_mut();
-        let entries = slf.entries.entry(entry.namespace_id()).or_default();
+        let entries = slf.entries.entry(*entry.namespace_id()).or_default();
         let new = entry.entry();
         let mut to_remove = vec![];
         for (i, existing) in entries.iter().enumerate() {
@@ -218,15 +218,15 @@ impl traits::EntryStorage for Rc<RefCell<EntryStore>> {
             if existing == new {
                 return Ok(false);
             }
-            if existing.subspace_id == new.subspace_id
-                && existing.path.is_prefix_of(&new.path)
+            if existing.subspace_id() == new.subspace_id()
+                && existing.path().is_prefix_of(&new.path())
                 && existing.is_newer_than(new)
             {
                 // we cannot insert the entry, a newer entry exists
                 return Ok(false);
             }
-            if new.subspace_id == existing.subspace_id
-                && new.path.is_prefix_of(&existing.path)
+            if new.subspace_id() == existing.subspace_id()
+                && new.path().is_prefix_of(&existing.path())
                 && new.is_newer_than(existing)
             {
                 to_remove.push(i);
@@ -258,15 +258,13 @@ impl CapsStore {
         // Select the best candidate, by sorting for
         // * first: widest area
         // * then: smallest number of delegations
-        let best = candidates.reduce(
-            |prev, next| {
-                if next.is_wider_than(prev) {
-                    next
-                } else {
-                    prev
-                }
-            },
-        );
+        let best = candidates.reduce(|prev, next| {
+            if is_wider_than(next, prev) {
+                next
+            } else {
+                prev
+            }
+        });
         Ok(best.cloned())
     }
 
@@ -281,7 +279,7 @@ impl CapsStore {
         // Select the best candidate, by sorting for
         // * widest area
         let best = candidates.reduce(|prev, next| {
-            if next.read_cap().is_wider_than(prev.read_cap()) {
+            if is_wider_than(next.read_cap(), prev.read_cap()) {
                 next
             } else {
                 prev
@@ -319,15 +317,15 @@ impl CapsStore {
         match cap {
             CapabilityPack::Read(cap) => {
                 self.read_caps
-                    .entry(cap.read_cap().granted_namespace().id())
+                    .entry(*cap.read_cap().granted_namespace())
                     .or_default()
-                    .push(cap);
+                    .push(cap.into());
             }
             CapabilityPack::Write(cap) => {
                 self.write_caps
-                    .entry(cap.granted_namespace().id())
+                    .entry(*cap.granted_namespace())
                     .or_default()
-                    .push(cap);
+                    .push(cap.into());
             }
         }
     }
